LPD-33410 Move ChartTooltip to out of metrics folder
