+++ modules/dxp/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/components/ChartTooltip.tsx
@@ -5,8 +5,8 @@
 
 import React from 'react';
 
-import {Alignments, Weights} from '../../types/global';
-import TooltipTemplate from '../TooltipTemplate';
+import {Alignments, Weights} from '../types/global';
+import TooltipTemplate from './TooltipTemplate';
 
 export type Column = {
 	align?: Alignments;
