+++ modules/dxp/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/components/interactions-by-page/InteractionsByPageChartTooltip.tsx
@@ -7,8 +7,8 @@ import React from 'react';
 
 import {Alignments, RangeSelectors, Weights} from '../../types/global';
 import {formatTooltipDate} from '../../utils/date';
+import ChartTooltip from '../ChartTooltip';
 import {IAccessibleTickProps} from '../metrics/AccessibleTick';
-import ChartTooltip from '../metrics/ChartTooltip';
 import {DotProps} from '../metrics/Dots';
 import {DataKey} from '../metrics/MetricsChart';
 import {FormattedData} from '../metrics/utils';
