+++ modules/dxp/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/components/visitors-behavior/VisitorsBehaviorChartTooltip.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 
 import {Alignments, Colors, RangeSelectors, Weights} from '../../types/global';
 import {formatTooltipDate} from '../../utils/date';
-import ChartTooltip from '../metrics/ChartTooltip';
+import ChartTooltip from '../ChartTooltip';
 import {CircleDot, PublishedVersionDot} from '../metrics/Dots';
 import {FormattedData} from '../metrics/utils';
 import {VisitorsBehaviorDataKey} from './utils';
